change to default build scripts to match ktls

+++ scripts/ktls/rebuild_ktls.sh
@@ -6,7 +6,11 @@ source $ROOT_DIR/scripts/async_libsrcs.source
 [ -z "$(lsmod | grep -e '\btls\b' )" ] && sudo modprobe tls
 
 #check for build
-[ ! -d "${KTLS_DIR}" ] || [ ! -f "${KTLS_DIR}/openssl-3.0.0.tar.gz" ] || [ ! -d "${KTLS_DIR}/openssl-3.0.0" ] || [ ! -f "${KTLS_DIR}/nginx-1.21.4.tar.gz" ] || [ ! -d "${KTLS_DIR}/nginx-1.21.4" ] || [ ! -f "${KTLS_DIR}/nginx_build/sbin/nginx" ] && echo "missing directory" && ${ROOT_DIR}/scripts/ktls/ktls_build.sh >/dev/null
+if [ ! -d "${KTLS_DIR}" ] || [ ! -f "${KTLS_DIR}/openssl-3.0.0.tar.gz" ] || [ ! -d "${KTLS_DIR}/openssl-3.0.0" ] || [ ! -f "${KTLS_DIR}/nginx-1.21.4.tar.gz" ] || [ ! -d "${KTLS_DIR}/nginx-1.21.4" ] || [ ! -f "${KTLS_DIR}/nginx_build/sbin/nginx" ]; then 
+	echo "KTLS Directory not setup ..." 
+	${ROOT_DIR}/scripts/ktls/ktls_build.sh >/dev/null
+else
+	sudo rm -rf ${KTLS_DIR}/nginx-1.21.4/Makefile ${KTLS_DIR}/nginx_build
+	${ROOT_DIR}/scripts/ktls/ktls_build.sh $*
+fi
 
-sudo rm -rf ${KTLS_DIR}/nginx-1.21.4/Makefile ${KTLS_DIR}/nginx_build
-${ROOT_DIR}/scripts/ktls/ktls_build.sh $*
